Lesson6

+++ src/main/java/lesson5/dto/Product.java
@@ -15,7 +15,8 @@ import lombok.With;
 public class Product {
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
+//    id - int
     @JsonProperty("title")
     private String title;
     @JsonProperty("price")
